fix: stuck turns after permission answers + session.list float timestamps
- mux reader now filters frames to the target session at the socket and
  gives critical frames (turn/end, approval/question requests, stream
  errors) blocking priority — a flood from other sessions (e.g. the Web
  GUI) previously filled the shared channel and dropped turn/end, leaving
  the engine stuck after answering a permission prompt (/stop ineffective
  while the turn never completes).
- events channel 256→1024 and non-critical EventText/Thinking/Tool events
  drop instead of stalling the reader.
- sessionSummary.updatedAt parsed as float64: dsh reports fractional
  millisecond timestamps (1786465628680.5632); int64 unmarshalling failed,
  breaking ValidateSessionID and /list (every message started a fresh
  session).
- regression tests: fractional updatedAt; turn/end survives a 4500-frame
  other-session flood.

+++ go/dsh/dsh.go
@@ -155,7 +155,7 @@ func (a *Agent) ListSessions(ctx context.Context) ([]core.AgentSessionInfo, erro
 		}
 		info := core.AgentSessionInfo{
 			ID:         s.SessionID,
-			ModifiedAt: time.UnixMilli(s.UpdatedAt),
+			ModifiedAt: time.UnixMilli(int64(s.UpdatedAt)),
 		}
 		if s.Projections != nil && s.Projections.Values.Title != nil {
 			info.Summary = truncate(*s.Projections.Values.Title, 60)
